🧪 Address PR comments: rename file, fix imports, bump docs version
